[DWARF] Generalize DWARFTypePrinter to a template class
